[api] refactoring: 'constants' module renamed to 'paths'
