+++ bauh/gems/arch/__init__.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-from bauh.api.constants import CACHE_PATH, CONFIG_PATH, TEMP_DIR
+from bauh.api.paths import CACHE_PATH, CONFIG_PATH, TEMP_DIR
 from bauh.commons import resource
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
