+++ bauh/view/core/tray_client.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from bauh.api.constants import CACHE_PATH
+from bauh.api.paths import CACHE_PATH
 
 TRAY_CHECK_FILE = '{}/notify_tray'.format(CACHE_PATH)  # it is a file that signals to the tray icon it should recheck for updates
